Revamping pre and post install scripts
The pre and post update process was too simple to work for us. This
commit replaces it with a more automated version. We need to specify
the individual steps between versions and the updater will figure out
which of those need to be called.

Related to https://github.com/KanoComputing/peldins/issues/735