Use isinstance() rather than type() for validation
When checking for validity of function parameters, comparing `type()` to
a desired data type will only return true if the type exactly matches,
derived types will fail the match. The check is only there to ensure
that the methods which are assumed to be present in subsequent calls are
actually members of the data type; derived types will have these methods
so should be allowed.

Pylint also complains that the comparison is an 'unidiomatic typecheck'.