Notification: Add flag to disable when the GUI is running
KanoComputing/peldins#1902
When the updater is in full-screen install mode, it will change its
status file to account for the progress made in the install. This,
however, generates notifications about the status of the process which
are visible above the GUI itself. Prevent this by providing a flag in
the status file which can disable these messages.