Notifications: Don't show on boot if a dialog appears
KanoComputing/peldins#1915
When booting, if the updater had to reboot to be able to continue, a
message is shown to prompt the user to carry on with the update. This
accompanies a status change which triggers a notification. If this is
the case, disable the notification so that two calls to action aren't
present.