Notification: Remove status file check on init
KanoComputing/peldins#1923
When the notification widget is initialised, the status file is checked
and, if appropriate, a notification may be triggered. This causes
unknown lock-ups when a user attempts to click the buttons on the
notification.

Instead, remove this call in the constructor and rely on the status file
monitor to trigger the notifications. The Updater boot check modifies
the file anyway so the notification is triggered but without the
lock-up.