Priority: Add ability to install urgent updates
KanoComputing/engage#62
Add awareness of priorities to the updater. Add the behaviour only for
the case of an urgent update.

If an urgent update is found then update the status file to set the
update type to urgent. This will make all subsequent phases in the
current update cycle urgent; this means that if both an urgent and
normal update are available, the urgent takes priority and only it is
installed rather than permitting a combination of the two to be dealt
with.