Priority: Fix adding of urgent suite to sources
KanoComputing/engage#62
When adding the new `release-urgent` suite to the local APT sources, a
`grep` is performed and if not found the source is added to the file.
This, however, fails to behave as expected when the string being
searched is multiline (as happens to be the case). Instead switch to use
`pcregrep` which supports multiline queries.