Priority: Fix notification muting during urgent updates
KanoComputing/engage#62
When muting the notifications during urgent updates, a flag is set in
the status file. This was being set in the wrong location which meant
that it was being set when there were no updates rather than when there
were important updates. Resolve this by moving it into the correct if
statement.