Boot: Don't confirm urgent updates at boot
When urgent updates are found but the install at shutdown fails, the
install will retry at boot. This shouldn't be negotiable and thus the
user shouldn't be offered the option to do this later. Also, the
current implementation will do the update and then launch the boot GUI
(which should have been overridden by the urgent flag) so ensure that
the updater closes after this.