Priority: Only run install on download if urgent
When running `kano-updater download`, it was performing a download and
then calling `run_install`, regardless of whether the update was urgent.
While `run_install` does have some checks for urgency, it will perform
the install if the state isn't in a downloaded state and so can run if
there are no updates available, causing a dialog to appear. Add a check
before running to ensure that this doesn't happen.