Priority: Change method of adding urgent repository
When adding the urgent repository to the sources, the original system
would just check for a specific string and, if it wasn't present, it
would directly add the string. Switch to a process which instead picks
out the Kano repository and appends the `release-urgent` source. This
solves the problem that the source which was added before would always
be the `repo.kano.me`.