Refactored updater restart/shutdown mechanism
Previously, a restart was triggered if the update had been scheduled. Since scheduling
now also happens for regular updates, the flag is no longer viable to distinguish between
the cases. A parameter was added for when `sudo kano-updater ui shutdown-window` is called
such that the finishing method is a shutdown. In all other cases, the kit will reboot.