Added Sounds to flappy-judoka and improved launch from updater
The Sound class now plays one fo the four sounds of the game. It implements various
listeners from the game logic to receive notification about when to play them. As
such, it is entirely decoupled from the game. A TODO would be to add a cmdline arg
to run the game without sounds if one needs.

The game launch from the updater was improved to only allow the game to run on RPi2.
This is due to performance issues of running it on top of the updater. However, the
game runs well by itself. Also, the launching mechanism was improved to only kill
necessary apps on the first launch. Subsequent game launches should now be faster.