Bluez: Move bluez preinstall to happen with every update
https://trello.com/c/1QrdXt2l/92-updater-pi2-update-from-2-4-0-to-3-4-0-failed
The update of `bluez` will often fail due to its postinstall
configuration script exiting with error status when the user does not
have a bluetooth capable device present. A work-around to this was
utilised for the update from v2.3.0 - v2.4.0 via a preupdate scenario
which installed the package manually. Unfortunately, this same problem
affects users updating from v2.4.0 to later versions so make the
solution more broad to do the install independently for every update.

This has a side-effect of adding a function where pre- and post-update
scripts which should run on every update can be located.