Debian: Require latest i18n versions of kano-toolset
Logging of errors can go awry if the version of kano-toolset which the
updater is started with (after updating itself) doesn't have the latest
i18n tools. This can cause logging of i18n related errors rather than
the original cause, making it very difficult to debug. Resolve this by
depending on the latest versions.