Download command now re-checks and re-downloads over an old cache
Previously, the Updater would not download newer packages in the
background once it had downloaded something once. Now, it will
check for updates again and when newer packages are found they
will be downloaded. This avoids installing an old set of packages
which could create problems over long periods of time.