Disk Space: Revert packages in reverse order
When checking for disk space requirements, every package is marked for
update and then reverted to its previous state after the space
requirement is calculated. Prevent reversion errors by restoring their
original state in the opposite order to that which set it in the first
place. Also, don't bother restoring package states which were already
marked for update before the whole process began.