Configure autologin during update installation
When the Updater is relauched at bootup to potentially continue an
interrupted update on a multi-user system, the Greeter will come
up and prevent the update from continuing until the a user logs in.
Perform an autologin on the user account that started the update
to continue the installation process.