Clean the apt cache after installation
It seems the apt cache is not cleaned completely with autoremove
after a download was completed. Upgrading an image from v2.4.0
shows some 500MB could be freed and autoclean is not helping with
that. Running the apt clean command after an update installation
should help with keeping the cache down reliably.