Autoexpand: Fix autoexpand script
With the introduction of the recovery partition, more needs to be done
to support auto-expanding the SD card as there are additional partitions
to contend with while still maintaining backwards-compatibility with the
standard image layout. To do this, split the expand out into a set of
Python functions to simplify.