Fix get_required_upgrade_space manual pkg marking and reverting
This issue presented itself when upgrading to Stretch, so it will
be required in an OS version before transitioning to Stretch. The
issue came about when manually marking and reverting pkg states
presumably in an order which caused a apt broken pkgs state. Use
the upgrade(dist_upgrade=True) and clear() functions of the
apt.cache to do the same safer and much much quicker.