Tests: Add tests for scenarios
Test that the scenarios correctly "cover" every update scenario to avoid
scenarios being missed when bumping the version. Fix a problem with the
scenarios which would incorrectly report allowed coverage between two
versions where the allowed update transitions cover more the requested
end version but don't have an explicit transition to the requested end
version.

Also add tests to the version to ensure that the tests are correctly
updated when bumping the version.