Stretch Update: Add upgrade path to v4.x.x
Addresses OS-29.

Up to now the Debian sources lists have been unmanaged and placed in the
system manually, migrate them into a package to allow them to be
properly managed, recoverable and verifiable.

Add a pre-update step to the updater to switch the sources over to the
new repository for those users who have hardware compatible with v4.x.x
and relaunch the updater flow once those are in place to trigger the
update. For those staying on Jessie, 3.16.0 will be the final update and
the updater will clean up to ensure that all is well.