Apt: Change update marking strategy
The scheme for marking packages for update would iterate through each
package and query if the package had an update available for it. This
could result in a package which has an update available being marked
while another, more essential package could conflict with the first and
result in a broken state. Resolve this by using the dist-upgrade instead
of manually iterating through packages.