Sources: Don't fail when failing to fetch sources
Failing to update sources isn't necessarily a cause for error - it is
only if the packages cannot subsequently be downloaded and installed is
there a problem so don't throw and error. This is even more pertinent
now that the development sources are present (if disabled) by default.

Additionally, disable Jessie i18n sources as they aren't correct and it
needs its own handling and QA anyway.