Cache: Add method to completely refresh the cache
When grabbing an instance to the apt wrapper handle, currently it will
always return the same cache instance and will never fully refresh. This
can cause a problem when the underlying sources change as it will not
re-read these when it updates. Add a method to refresh this and utilise
for the Stretch upgrade.