Tests: Expect install test failure when relaunch required
The updater relaunches when the system version is less than the target
version. With the changes to prevent the update continuing after an
error, the install function returns with a fail return value now in this
scenario so the relaunch is coupled to the success of the install. Mark
these specific test cases as expected failures.