Removed v3.7 to v3.8 3P app installation disk space requirement
Before, the disk space requirement was part of the update itself.
Moved away from that model to install apps separately of the update
to minimise the requirement (divide and conquer).
However, the App Store does not check for disk space before
installing an app so account for this here naively.
Requirements (disk_req) in MB were made with apt on a clean system.