COPPA: Set the parental setting in the background
During the post-update scenarios the parental setting is set to be COPPA
compliant. When we do this, the `kano-settings-cli` command it used but
this blocks while the sentry proxy server is restarted and then owns the
server itself and thus never returns, causing a hang. To resolve this,
trigger the parental setting configuration in the background and wait
for the process to establish itself.